Implement Add Treatment command
Pets usually have treatments.

This command so that users can add treatment records for
their pets using the index of pet, name of treatment, date of treatment,
and type of treatment.

Let's add a script to enable users to add treatment records.

This implementation validates index and date format and prints
user-friendly errors on invalid input.