Implement logging and assertions for add-pet command
Currently, there is no safety net for the add-pet command
in case the user decides to input invalid information. In
other words, there is no add-pet log to access critical information
to troubleshoot any problems within the code base.

With this feature, the add-pet command now has troubleshooting
capabilities and ensures that the application is more defensive.