Add developer guide for AddPetCommand
Currently, there is no proficient documentation for the
AddPetCommand. If the user or other engineers want to
know more about the workflow and usage of the command,
they would not be able to learn without the user guide.

With this addition, there is simple and easy-to-follow
documentation about the add-pet command including
the workflow within the application as a whole.